Most of task one working....
So, i have got the object to move side to side but Im having trouble
with the rotation. It wont rotate around the origin of the object and so
I have tried translating it to take this into account and then rotating
then translating back but that has made no difference. I am currently
using in buildt mat4 functions(there is an api online if you dont
understand the functions) to do the rotations and translation but I have
also tried it with matrix operations in the commented out section of the
code. I will probably keep working on it today but that is where I am up
to